requirements.txt

pandas==2.2.2
numpy==1.26.4
pandas-ta==0.3.14b0
polars>=1.4.0
pyarrow>=16.0
XlsxWriter>=3.2
openpyxl>=3.1
click>=8.1
tqdm>=4.66
pydantic>=2.7
loguru>=0.7
PyYAML>=6.0
pytest>=8.2
